requirements.txt
click==8.1.7
ImageHash==4.3.1
ipykernel==6.27.1
ipython==8.14.0
joblib==1.3.1
lpips==0.1.4
matplotlib==3.8.2
numpy==1.24.3
pandas==1.5.3
Pillow==10.1.0
pyarrow==12.0.1
pytorch-lightning==2.0.5
scikit-learn==1.3.0
sentence-transformers==2.5.1
torch==2.0.1
torchsampler==0.1.2
torchvision==0.15.2
tqdm==4.65.0
umap-learn==0.5.5